requirements.txt
av==9.2.0
certifi==2023.11.17
chumpy==0.70
Cython==3.0.8
easydict==1.12
einops==0.7.0
face-alignment==1.4.1
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.3
fvcore==0.1.5.post20221221
gpustat==1.1.1
huggingface-hub==0.0.8
imageio==2.33.1
imageio-ffmpeg==0.4.9
iopath==0.1.10
jiwer==2.3.0
keras==2.15.0
kornia==0.6.6
librosa==0.9.2
lmdb==1.4.1
loguru==0.6.0
matplotlib==3.8.3
mediapipe==0.8.10
multiprocess==0.70.15
ninja==1.11.1.1
numba==0.58.1
numpy==1.23.0
omegaconf==2.1.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.1
phonemizer==3.2.1
pillow==10.2.0
protobuf==3.20.3
psbody-mesh==0.4
pyglet==2.0.12
PyOpenGL==3.1.0
pyrender==0.1.45
PyYAML==6.0.1
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.12.0
six==1.16.0
soundfile==0.12.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0.post1
tokenizers==0.10.3
torchfile==0.1.0
torchgeometry==0.1.2
torchmetrics==1.3.0
tqdm==4.66.1
transformers==4.6.1
trimesh==3.16.4
yacs==0.1.8